Sort list of links by dragging them (#20)
* Install react-drag-reorder

* Add Draggable component

* Replace ul element with Draggable component

* Restore LinkedListItem component functionality
Accidentally removed the functionality for this component
while working on updating the styling. Will double-check PR
changes before merging anything.

* Add initial functionality that runs when items change position

* Add react-easy-sort

* Remove Draggable component

* Replace li with div element

* Add key to mapped link list item

* Replace ul with div element

* Add LinkList component

* Immutably update links array when items are switched

* Add LinkList component to App

* Remove console log

* Remove LinkListItem component

The package that adds the sorting functionality doesn't
work with components directly. While the documentation says
it is possible by using the forwardRef() function, it did not
seem to work. The easiest workaround was to copy the component
code into the LinkList component.

* Remove react-drag-reorder